+++ CK.MQTT.Client/Mqtt3SinkWrapper.cs
@@ -31,7 +31,7 @@ namespace CK.MQTT.Client
 
         protected abstract void OnPoisonousPacket( ushort packetId, PacketType packetType, int poisonousTotalCount );
 
-        protected abstract void OnPacketResent( ushort packetId, int resentCount, bool isDropped );
+        protected abstract void OnPacketResent( ushort packetId, ulong resentCount, bool isDropped );
 
         protected virtual void OnQueueFullPacketDropped( ushort packetId, PacketType packetType ) { }
         protected virtual void OnQueueFullPacketDropped( ushort packetId ) { }
@@ -53,7 +53,7 @@ namespace CK.MQTT.Client
 
         void IMqtt3Sink.Connected() => OnConnected();
 
-        void IMqtt3Sink.OnPacketResent( ushort packetId, int resentCount, bool isDropped ) => OnPacketResent( packetId, resentCount, isDropped );
+        void IMqtt3Sink.OnPacketResent( ushort packetId, ulong resentCount, bool isDropped ) => OnPacketResent( packetId, resentCount, isDropped );
 
         void IMqtt3Sink.OnQueueFullPacketDropped( ushort packetId, PacketType packetType ) => OnQueueFullPacketDropped( packetId, packetType );
         void IMqtt3Sink.OnQueueFullPacketDropped( ushort packetId ) => OnQueueFullPacketDropped( packetId );
